requirements.txt
pytorch==1.11.0+cu113
transformers==4.27.4
pytorch-crf==0.7.2
datasets
tokenizers